Improve bash file to create symbolic links for config files

+++ copy_config.sh
@@ -1,13 +1,25 @@
 #!/bin/bash
 
-echo "Backup and the copy contents of bashrc to home folder."
-cp ~/.bashrc ~/".bashrc_$(date +%F_%R)"
-cat ./.bashrc > ~/.bashrc
+dir=`pwd`
 
-echo "Backup and the copy vim config to home folder."
-cp ~/.vimrc ~/".vimrc_$(date +%F_%R)"
-cat ./.vimrc > ~/.vimrc
+echo "Backup and create link for vimrc file"
 
-echo "Append gitconfig to git configuration file. Remove any extra lines."
-cat ./.gitconfig >> ~/.gitconfig
-vi ~/.gitconfig
+if [ -f ~/.vimrc ]; then
+  cp ~/.vimrc ~/".vimrc_$(date +%F_%R)"
+  rm ~/.vimrc
+fi
+
+if [ ! -L ~/.vimrc ]; then
+  ln -s $dir/.vimrc ~/.vimrc
+fi
+
+echo "Backup and the copy alias file to home folder."
+
+if [ -f ~/.aliases ]; then
+  cp ~/.aliases ~/".aliases_$(date +%F_%R)"
+  rm ~/.aliases
+fi
+
+if [ ! -L ~/.aliases ]; then
+  ln -s $dir/.aliases ~/.aliases
+fi
